- [New] started creating the prop system - [Rewrite] Started full rewrite of the animation system again. We now use quaternions for rotations (again) and transformation matrices around that. It is nearly done except for the following issues:   - bones that have a default rotation large enough to change the axis of rotation of the animation are not getting rotated in the right direction due to the minecraft models being offset by 180 degrees compared to the expectation   - the z rotation being inverted from the usual rotation system is causing problems for the local rotations in the animations

+++ src/main/java/com/magmaguy/freeminecraftmodels/dataconverter/AnimationFrame.java
@@ -1,15 +1,15 @@
 package com.magmaguy.freeminecraftmodels.dataconverter;
 
 public class AnimationFrame {
-    public double xRotation;
-    public double yRotation;
-    public double zRotation;
-    public double xPosition;
-    public double yPosition;
-    public double zPosition;
-    public double xScale;
-    public double yScale;
-    public double zScale;
+    public float xRotation;
+    public float yRotation;
+    public float zRotation;
+    public float xPosition;
+    public float yPosition;
+    public float zPosition;
+    public float xScale;
+    public float yScale;
+    public float zScale;
 
     public AnimationFrame(){
     }
